Tethering script from Space Marooner
Script that manages the "force field" that keeps the player near their ship. When nearing the edge of the tethering length, the player is met with an impassable force field.

This is ofcourse just smoke screens and a trick of the eye, as the force field is but a mere quad that follows the players position when near the edge and has a very convincing shader to fool any player into thinking the whole ship is surrounded by a visible bubble. As the player builds their ship, the force field grows according to the building direction.

This allows the player to build as much as they please without the tethering tool ever getting in the way of their creativity